requirements.txt
# OpenClaw — Agentic Automation Platform
# Core LLM
mistralai>=1.0.0

# Gmail / Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Web Search
tavily-python>=0.3.0

# HTTP / Orchestrator communication
flask>=3.0.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
# Streamlit UI
streamlit>=1.35.0
